+++ apps/cubics-campaigns/app/layout.tsx
@@ -1,8 +1,9 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import { SessionProvider } from "next-auth/react";
-import "./globals.scss";
 import Header from "@/components/Header/header";
+import "@studiocubics/styles/cubicscampaigns";
+import "@studiocubics/core/styles";
 
 const inter = Inter({ subsets: ["latin"] });