+++ packages/core/src/components/index.ts
@@ -23,5 +23,3 @@ export { default as Divider } from "./Divider/Divider";
 export { default as DragNDrop } from "./DragNDrop/DragNDrop";
 
 export { default as LoadingIcon } from "./LoadingIcon/LoadingIcon";
-
-export { default as Toast } from "./Toast/Toast";