+++ packages/icons/rollup.config.js
@@ -1,5 +1,2 @@
 const common = require("@studiocubics/rollup-config/library.js");
-module.exports = {
-  external: ["react", "@react-email/components", "react-hot-toast"], // Exclude deps from the bundle
-  ...common,
-};
+module.exports = common;