+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @studiocubics/core
 
+## 1.2.0
+
+### Minor Changes
+
+- minor fixes and changes to exports
+
+## 1.1.0
+
+### Minor Changes
+
+- styles package added to added default styles to webapps.
+
 ## 1.0.0
 
 ### Major Changes