+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @studiocubics/ui
 
+## 1.2.0
+
+### Minor Changes
+
+- minor fixes and changes to exports
+
+### Patch Changes
+
+- Updated dependencies
+  - @studiocubics/email@1.2.0
+  - @studiocubics/icons@1.2.0
+  - @studiocubics/core@1.2.0
+
+## 1.1.0
+
+### Minor Changes
+
+- styles package added to added default styles to webapps.
+
+### Patch Changes
+
+- Updated dependencies
+  - @studiocubics/email@1.1.0
+  - @studiocubics/icons@1.1.0
+  - @studiocubics/core@1.1.0
+
 ## 1.0.0
 
 ### Major Changes